refactor: consolidate searchable language grid and remove legacy templates
