--- app/eventyay/control/templates/pretixcontrol/multi_languages_widget.html
@@ -1,9 +1,0 @@
-{% load i18n %}
-{% if widget.wrap_label %}
-    <label{% if widget.attrs.id %} for="{{ widget.attrs.id }}"{% endif %}>
-{% endif %}
-{% include "django/forms/widgets/input.html" %}
-{% if widget.wrap_label %}
-    {{ widget.label }}
-    </label>
-{% endif %}
