+++ app/eventyay/orga/forms/widgets.py
@@ -1,30 +1,9 @@
 from django.conf import settings
-from django.forms import CheckboxSelectMultiple, RadioSelect
+from django.forms import RadioSelect
 
 
 class HeaderSelect(RadioSelect):
     option_template_name = 'orga/widgets/header_option.html'
 
 
-class MultipleLanguagesWidget(CheckboxSelectMultiple):
-    template_name = 'orga/widgets/language_grid_select.html'
-    option_template_name = 'orga/widgets/multi_languages_widget.html'
 
-    def __init__(self, *args, **kwargs):
-        kwargs['attrs'] = kwargs.get('attrs', {})
-        kwargs['attrs']['class'] = kwargs['attrs'].get('class', '') + ' multi-language-select'
-        kwargs['attrs']['help_left'] = True
-        super().__init__(*args, **kwargs)
-
-    def _sorted_choices(self):
-        # Sort by translated language names in the current UI locale.
-        choices = list(self.choices)
-        translated_names = {code: str(name) for code, name in settings.LANGUAGES}
-        return sorted(
-            choices,
-            key=lambda c: (str(translated_names.get(c[0], c[1])).casefold(), str(c[0])),
-        )
-
-    def optgroups(self, name, value, attrs=None):
-        self.choices = self._sorted_choices()
-        return super().optgroups(name, value, attrs)
