+++ app/eventyay/orga/templates/orga/base.html
@@ -51,7 +51,6 @@
             <link rel="stylesheet" type="text/css" href="{% static 'orga/css/_flags.css' %}" />
             <link rel="stylesheet" type="text/css" href="{% static 'orga/css/_layout.css' %}" />
             <link rel="stylesheet" type="text/css" href="{% static 'common/css/_component_nav_mobile.css' %}" />
-            <link rel="stylesheet" type="text/css" href="{% static 'eventyay-common/css/language-grid.css' %}" />
             <link rel="stylesheet" type="text/css" href="{% static 'orga/css/_rtl.css' %}" />
         {% endcompress %}
         {% if request.event and request.event.visible_primary_color %}<link rel="stylesheet" type="text/css" href="{{ request.event.urls.settings_css }}?target=orga&c={{ request.event.visible_primary_color|urlencode }}" />{% endif %}
@@ -72,7 +71,6 @@
             <script defer src="{% static 'orga/js/sidebar.js' %}"></script>
             <script type="module" src="{% static 'eventyay-common/js/ajax-filter.js' %}"></script>
         {% endcompress %}
-        <script defer src="{% static 'eventyay-common/js/language-grid.js' %}"></script>
         {% comment %}
         This JS file is generated by Vue 3 app. Don't compress it with Django Compressor. The output is conflicted with other scripts.
         {% endcomment %}
