--- app/eventyay/orga/templates/orga/widgets/multi_languages_widget.html
@@ -1,12 +1,0 @@
-{% load i18n %}
-{% include "django/forms/widgets/input.html" %}
-{% if widget.wrap_label %}
-  <label class="form-check-label" {% if widget.attrs.id %}for="{{ widget.attrs.id }}"{% endif %}>
-    {% if not widget.attrs.hide_label %}{{ widget.label }}{% endif %}
-    {% if not widget.official and widget.percentage %}
-      <a href="#">
-        <span class="badge badge-pill color-{% if widget.percentage >= 90 %}success{% elif widget.percentage >= 80 %}warning{% else %}danger{% endif %}">
-          {% blocktranslate trimmed with percentage=widget.percentage %}{{ percentage }} % translated{% endblocktranslate %}</span></a>
-    {% endif %}
-  </label>
-{% endif %}
